# backend-collect/doc/api.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Terminator111/COLLECT/5.0.0
info:
  description: 软工三项目接口文档
  version: 4.0.0
  title: COLLECT-API
...
paths:
...
  /admin/task/all:
    get:
      tags:
        - admin
      summary: 管理员预览任务
      description: 管理员查看所有任务列表（预览）
      parameters:
        - name: pageId
          in: query
          schema:
            type: integer
          required: true
        # - name: sortType
        #   in: query
        #   schema:
        #     type: integer
        #   required: true
        #   description: 0代表按发表时间倒序，1代表按最新回复时间倒序
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskViewVO"
  /admin/task/detail/{taskId}:
    get:
      tags:
        - admin
      summary: 管理员查看任务详细信息
      description: 管理员从任务列表中点击某一个任务，查看任务详细信息
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/TaskVO'
...
  /admin/recommendRule/delete/{ruleId}:
    delete:
      tags:
        - admin
      summary: 管理员删除推荐规则
      description: 管理员删除一个推荐规则，要确保管理员没有删除正在使用的推荐规则
      parameters:
        - name: ruleId
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVO'
  /admin/recommendRule/add:
    post:
      tags:
        - admin
      summary: 管理员添加推荐规则
      description: 管理员向推荐列表中添加一条推荐规则
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRuleDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecommendRuleVO'
  #  /admin/recommendRule/edit/{ruleId}:
  #    post:
  #      tags:
  #        - admin
  #      summary: 管理员修改一条推荐规则
  #      description: 管理员修改推荐规则列表中的一条推荐规则
  #      parameters:
  #        - name: ruleId
  #          schema:
  #            type: integer
  #          in: path
  #          required: true
  #      requestBody:
  #        content:
  #          application/json:
  #            schema:
  #              $ref: '#/components/schemas/RecommendRuleDTO'
  #      responses:
  #        '200':
  #          description: OK
  #          content:
  #            application/json:
  #              schema:
  #                type: object
  #                properties:
  #                  code:
  #                    type: integer
  #                  msg:
  #                    type: string
  #                  data:
  #                    type: array
  #                    items:
  #                      $ref: '#/components/schemas/RecommendRuleVO'
  /admin/recommendRule/select/{ruleId}:
    post:
      tags:
        - admin
      summary: 管理员启用推荐规则
      description: 管理员在推荐规则列表，选择启用某一条推荐规则（仅一条为启用状态）
      parameters:
        - name: ruleId
          in: path
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseVO"
...
  /task/detail/{taskId}:
    get:
      tags:
        - task
      summary: 查看任务详细信息
      description: 发包方/众包工人从任务列表中点击某一个任务，查看任务详细信息
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/TaskVO"
...
  /flaw/refine/{flawId}:
    post:
      tags:
        - flaw
      summary: 完善一个缺陷
      description: 众包工人可以在报告完善过程中选择fork一个其他人的缺陷, 请求体中的内容是用户修改后的缺陷
      parameters:
        - name: flawId
          description: 当前缺陷的Id
          in: path
          schema:
            type: integer
          required: true
        - name: forkedFlawId
          description: 被fork的缺陷的id
          in: query
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlawDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVO'
  /flaw/nofork/{flawId}: #这个接口会不会有点多余？
    get:
      tags:
        - flaw
      summary: 不fork任何一个缺陷
      description: 用户认为系统推荐的相似缺陷中没有任何一个缺陷和自己的缺陷相同，决定直接提交自己的缺陷而不去fork任何缺陷。
      parameters:
        - name: flawId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVO'
  /flaw/tree/{flawId}:
    get:
      tags:
        - flaw
      summary: 获取与该缺陷相关联的缺陷树（暂定）
      description: 获取与该缺陷相关的缺陷树
      parameters:
        - name: flawId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/FlawTreeNodeVO'
    # /flaw/tree/{flawId}:
    # get:
    #   tags:
    #   -flaw
    #   summary: 获取与该缺陷相关联的缺陷树
    #   description: 获取与该缺陷相关的缺陷树, 只返回当前节点及其子节点
    #   parameters:
    #     - name: flawId
    #       in: path
    #       schema:
    #         type: integer
    #       required: true
    #   responses:
    #     '200':
    #       description: OK
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schema/FlawTreeNodeVO'
  /flaw/similarFlaw/{flawId}:
    get:
      tags:
        - flaw
      summary: 获取与当前缺陷相似的缺陷列表
      description: 获取与当前缺陷相似的缺陷列表（按照相似度从高到低排序）
      parameters:
        - name: flawId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SimilarFlawVO'
  /flaw/myScore/{flawId}:
    get:
      tags:
        - flaw
      summary: 获取某用户对该缺陷的评分
      description: 获取某用户对该缺陷的评分
      parameters:
        - name: flawId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: integer
                    description: 用户对该缺陷的评分值，如果为-1表示未评分
  /flaw/score/:
    post:
      tags:
        - flaw
      summary: 对缺陷进行评分
      description: 对某个缺陷进行评分
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlawScoreDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVO'
  /flaw/flawMap/{taskId}:
    get:
      tags:
        - flaw
      summary: 获取某个任务的缺陷图
      description: 缺陷图中的每个节点都是某棵缺陷树的根节点，也就是说图中的每个节点都没有fork任何其他的缺陷, 对数据进行归一化和去中心化所有图中节点的均值在(0, 0)并且所有点的横纵坐标都在[-1,1]范围内
      parameters:
        - name: taskId
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/FlawMapVO'
  #  /flaw/flawSimilarity/{flawId}:
  #    post:
  #      tags:
  #        - flaw
  #      summary: 获取给定缺陷和多个缺陷的相似度
  #      description: 给定一个由缺陷id所组成的数组，获取id为flawId的缺陷与数组中所有缺陷的相似度，按照数组中的顺序返回
  #      parameters:
  #        - name: flawId
  #          in: path
  #          schema:
  #            type: integer
  #          required: true
  #      requestBody:
  #        required: true
  #        content:
  #          application/json:
  #            schema:
  #              type: array
  #              items:
  #                type: integer
  #      responses:
  #        '200':
  #          description: OK
  #          content:
  #            application/json:
  #              schema:
  #                type: object
  #                properties:
  #                  code:
  #                    type: integer
  #                  msg:
  #                    type: string
  #                  data:
  #                    type: array
  #                    items:
  #                      type: number
  #                      maximum: 1.0
  #                      minimum: 0.0
  #  /tag/groups:
  #    get:
  #      tags:
  #        - tag
  #      summary: 获取标签组（暂定）
  #      description: 获取所有的标签组，作为选项，这些标签可能被用在任务发布或者众包工人注册信息上
  #      responses:
  #        '200':
  #          description: OK
  #          content:
  #            application/json:
  #              schema:
  #                type: object
  #                properties:
  #                  code:
  #                    type: integer
  #                  msg:
  #                    type: string
  #                  data:
  #                    type: array
  #                    items:
  #                      $ref: '#/components/schemas/TagGroupVO'
  /flaw/evaluations/{flawId}:
    get:
      tags:
        - flaw
      summary: 获取缺陷的所有评价
      parameters:
        - name: flawId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          description: 当前登录用户的id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FlawEvaluationVO'
  /flaw/add/evaluation/{flawId}:
    post:
      tags:
        - flaw
      summary: 给非自己的报告中的缺陷增加一条评价
      parameters:
        - name: flawId
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlawEvaluationDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FlawEvaluationVO'
  /flaw/evaluation/like/{evaluationId}:
    post:
      tags:
        - flaw
      summary: 给某个评论点赞，默认可以点赞
      parameters:
        - name: evaluationId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/FlawEvaluationVO'
  /flaw/evaluation/dislike/{evaluationId}:
    post:
      tags:
        - flaw
      summary: 给某个评论点踩，默认可以点踩
      parameters:
        - name: evaluationId
          in: path
          schema:
            type: integer
          required: true
        - name: userId
          in: query
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  msg:
                    type: string
                  data:
                    $ref: '#/components/schemas/FlawEvaluationVO'
  /flaw/add/append:
    post:
      tags:
        - flaw
      summary: 增加或修改补充内容
      description: 无论是增加还是修改补充内容都调用这一个接口，后端直接覆盖字符串即可，只能对自己的报告增加/修改补充内容
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlawAppendDTO"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseVO"
...
components:
  schemas:
    ResponseVO:
      type: object
      description: api返回格式
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          type: object
...
    UserViewVO:
      type: object
      description: 用户简略个人信息展示
      properties:
        id:
          type: integer
        type:
          type: integer
          description: 0代表发包方，1代表众包工人，2代表管理员（注册的时候前端只有0，1两个选项）
        username:
          type: string
...
    TaskVO:
      type: object
      description: 众测任务
      properties:
        id:
          type: integer
        userId:
          type: integer
          description: 发包方的id
        type:
          type: integer
          description: 0代表功能测试，1代表性能测试
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerNum:
          type: integer
        # others……
        pickedWorkerNum:
          type: integer
        exeFileName:
          type: string
        reqDocName:
          type: string
        state:
          type: integer
          description: 用来描述任务的状态。0表示未领取或者请求该任务详情的用户不是众包工人，1表示进行中，2表示已完成，3表示逾期未交，4表示待完善（缺陷没有完全被处理），5表示任务已停止招募
        createTime:
          type: string
          format: date-time
        tagGroups:
          type: array
          items:
            type: integer
    TaskViewVO:
      type: object
      description: 任务预览信息
      properties:
        id:
          type: integer
        userId:
          type: integer
          description: 发包方的id
        type:
          type: integer
          description: 0代表功能测试，1代表性能测试
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerNum:
          type: integer
        pickedWorkerNum:
          type: integer
        tagGroups:
          type: array
          items:
            type: integer
...
    FlawDTO:
      type: object
      description: 缺陷
      properties:
        pictureURLs:
          type: array
          items:
            type: string
        description:
          type: string
        stepDes:
          type: string
        deviceInfo:
          type: string
    ## change3: 我添加了一个flawVO类型，相比原来的flawVO，这个类型中多出了score和forkedId属性
    FlawEvaluationDTO:
      type: object
      description: 缺陷评价
      properties:
        userId:
          type: integer
        content:
          type: string
    FlawAppendDTO:
      type: object
      description: 缺陷补充
      properties:
        flawId:
          type: integer
        appendContent:
          type: string
    FlawScoreDTO:
      type: object
      description: 缺陷评分
      properties:
        userId:
          type: integer
        flawId:
          type: integer
        score:
          type: integer
    FlawVO:
      type: object
      description: 缺陷
      properties:
        id:
          type: integer
        reportId:
          type: integer
        pictureURLs:
          type: array
          items:
            type: string
        description:
          type: string
        stepDes:
          type: string
        deviceInfo:
          type: string
        score:
          type: number
        scoreNum:
          type: integer
          description: 评分人数
        state:
          type: boolean
          description: 该缺陷是否已经被处理
        appendContent:
          type: string
          description: 对已提交的报告的补充内容
    FlawTreeNodeVO:
      type: object
      description: 缺陷树节点
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/FlawTreeNodeVO'
        current:
          $ref: '#/components/schemas/FlawVO'
    # FlawTreeNodeVO:
    #   type: object
    #   description: 缺陷树节点
    #   properties:
    #     children:
    #       type: array
    #       items: integer
    #     current:
    #       $ref: '#/components/schemas/FlawVO'
    SimilarFlawVO:
      type: object
      description: 相似缺陷
      properties:
        similarity:
          type: number
          maximum: 1.0
          minimum: 0.0
        flaw:
          $ref: '#/components/schemas/FlawVO'
...
    FlawEvaluationVO:
      type: object
      description: 某缺陷的单个评价
      properties:
        evaluationId:
          type: integer
        user:
          $ref: '#/components/schemas/UserViewVO'
        content:
          type: string
        likeInfo:
          $ref: '#/components/schemas/FlawEvaLikeVO'
    RecommendRuleDTO:
      type: object
      description: 推荐规则
      properties:
        name:
          type: string
        factors:
          type: array
          items:
            $ref: '#/components/schemas/RecommendRuleFactorVO'
    RecommendRuleVO:
      type: object
      description: 推荐规则
      properties:
        id:
          type: integer
        name:
          type: string
        using:
          type: boolean
        factors:
          type: array
          items:
            $ref: '#/components/schemas/RecommendRuleFactorVO'
    RecommendRuleFactorVO:
      type: object
      description: 推荐优先级影响因素
      properties:
        name:
          type: string
        weight:
          type: number
          maximum: 1.0
          minimum: 0.0
    FlawMapVO:
      type: object
      description: 缺陷图
      properties:
        flawIds:
          description: 根节点的flawId的列表
          type: array
          items:
            type: integer
        similarityMatrix:
          description: 矩阵中i行j列的元素代表flawIds中索引为i和j的缺陷的相似度
          type: array
          items:
            type: array
            items:
              type: number
    FlawEvaLikeVO:
      type: object
      description: 返回当前单个评论的点赞与点踩的数量，和当前用户对此条评论点赞点踩的状态（0—未点赞点踩，1-已点赞未点踩，2-已点踩未点赞，3-不能点赞和点踩）
      properties:
        likeNum:
          description: 当前单个评论的点赞数量
          type: integer
        dislikeNum:
          description: 当前单个评论的点踩数量
          type: integer
        status:
          description: 当前用户对此条评论点赞点踩的状态
          type: integer